API/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
eval_type_backport==0.2.0
faiss-cpu==1.9.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
grpcio==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpath-python==1.0.6
MarkupSafe==3.0.2
mistralai==1.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
portalocker==2.10.1
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.8.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
